Refactor tmdb -> api
