+++ medusa/indexers/config.py
@@ -5,7 +5,7 @@
 import re
 
 from medusa.app import BASE_PYMEDUSA_URL
-from medusa.indexers.tmdb.tmdb import Tmdb
+from medusa.indexers.tmdb.api import Tmdb
 from medusa.indexers.tvdb.api import TVDBv2
 from medusa.indexers.tvmaze.api import TVmaze
 from medusa.session.core import MedusaSession
